#116 -1のIdを定数化する
- -1をシステム値として定数化
- Idとして-1を指定不可に変更

+++ HouseholdAccountBook/ViewModels/Settings/BookSettingViewModel.cs
@@ -149,7 +149,7 @@ namespace HouseholdAccountBook.ViewModels.Settings
         /// <summary>
         /// 関連性セレクタVM
         /// </summary>
-        public SelectorViewModel<RelationViewModel, ItemIdObj> RelationSelectorVM { get; } = new(static vm => (int)vm?.Id);
+        public SelectorViewModel<RelationViewModel, ItemIdObj> RelationSelectorVM { get; } = new(static vm => vm?.Id);
         #endregion
 
         /// <summary>
